fix api

+++ AcessibilidadeWebAPI/Dtos/Voluntario/VoluntarioDto.cs
@@ -4,6 +4,6 @@
     {
         public int IdUsuario { get; set; }
         public bool Disponivel { get; set; }
-        public int Avaliacao { get; set; }
+        public decimal Avaliacao { get; set; }
     }
 }
